Rename companion objects back to non-default version
