+++ zmessaging/src/main/scala/com/waz/api/impl/Avs.scala
@@ -19,8 +19,8 @@ package com.waz.api.impl
 
 import com.waz.ZLog._
 import com.waz.api.AvsLogLevel
-import com.waz.service.call.{DefaultFlowManagerService}
-import com.waz.service.call.DefaultFlowManagerService.AvsLogData
+import com.waz.service.call.DefaultFlowManagerService
+import com.waz.service.call.FlowManagerService.AvsLogData
 import com.waz.threading.Threading
 import com.waz.ui.{SignalLoading, UiModule}
