+++ zmessaging/src/main/scala/com/waz/service/call/FlowManagerService.scala
@@ -21,6 +21,7 @@ import android.content.Context
 import android.view.View
 import com.waz.HockeyApp
 import com.waz.ZLog._
+import com.waz.ZLog.ImplicitTag._
 import com.waz.api._
 import com.waz.call._
 import com.waz.log.LogHandler
@@ -48,7 +49,7 @@ trait FlowManagerService {
 }
 
 class DefaultFlowManagerService(context: Context, netClient: ZNetClient, websocket: WebSocketClientService, prefs: PreferenceService, network: DefaultNetworkModeService) extends FlowManagerService {
-  import DefaultFlowManagerService._
+  import FlowManagerService._
 
   val MetricsUrlRE = "/conversations/([a-z0-9-]*)/call/metrics/complete".r
   val avsAudioTestFlag: Long = 1 << 1
@@ -330,8 +331,7 @@ case class AvsMetrics(rConvId: RConvId, private val bytes: Array[Byte]) {
   override def toString = s"AvsMetrics($rConvId, isVideoCall: $isVideoCall, kindOfCall: $kindOfCall, ${json.toString})"
 }
 
-object DefaultFlowManagerService {
-  private implicit val logTag: LogTag = logTagFor(DefaultFlowManagerService)
+object FlowManagerService {
 
   case class AvsLogData(metricsEnabled: Boolean, loggingEnabled: Boolean, logLevel: AvsLogLevel)
   case class EstablishedFlows(convId: RConvId, users: Set[UserId])
