+++ zmessaging/src/main/scala/com/waz/service/call/VoiceChannelService.scala
@@ -25,7 +25,7 @@ import com.waz.model.Event.CallProperties
 import com.waz.model.VoiceChannelData.ConnectionState
 import com.waz.model._
 import com.waz.service._
-import com.waz.service.call.DefaultFlowManagerService.EstablishedFlows
+import com.waz.service.call.FlowManagerService.EstablishedFlows
 import com.waz.service.conversation.ConversationsContentUpdater
 import com.waz.service.push.PushService
 import com.waz.sync.SyncServiceHandle
